修正当__MSVCRT_VERSION__ < 0x0800时读取和生成ParserTable不一致的Bug(将sizeof(time_t)对齐到8字节)
git-svn-id: http://qlanguage.googlecode.com/svn/trunk@242 6d982bb3-a5a7-393d-ffc5-582415be9c85

+++ Source/QLanguage/Parser/LALR1ParserTable.cpp
@@ -173,7 +173,8 @@ namespace QLanguage
         time(&t);
         uchar version = 1;
         fstream stream(path, fstream::out | fstream::binary);
-        stream << PARSER_TABLE << version << t << vts.size() << vns.size() << _rules.size() << items.size() << this->pStart->idx;
+        // Align time_t with 4 bytes to size 8 bytes
+        stream << PARSER_TABLE << version << (__int64)t << vts.size() << vns.size() << _rules.size() << items.size() << this->pStart->idx;
         for (vector<pair<uchar, ushort> >::const_iterator i = table.begin(), m = table.end(); i != m; ++i)
         {
             stream << i->first << i->second;
